log in/log out complete
users can now log in and log out. Both user@name.com:password and
c@d.com:dddddd work to log in after hitting the menu item button. Input
type for the email section was changed to TextEmail so as to add the "@"
and ".com" buttons to the keyboard.  There is only one menu item that
changes functionality and title to be used as both the log in and log
out button. After hitting the log in or log out menu item, there will be
a toast that pops up that lets the user know what the status of their
session is. The logout() function is called in onCreate in
MainActivity... but it might not need to be there.